Move gigglebot drivers to their own directory
